# perf: Reduce allocations