Before
test bench_parse_template  ... bench:      25,960 ns/iter (+/- 2,495)
test bench_parse_text      ... bench:         854 ns/iter (+/- 60)
test bench_parse_variable  ... bench:       2,802 ns/iter (+/- 241)
test bench_render_template ... bench:       8,106 ns/iter (+/- 578)
test bench_render_text     ... bench:         456 ns/iter (+/- 29)
test bench_render_variable ... bench:       1,617 ns/iter (+/- 117)

After:
test bench_parse_template  ... bench:      25,607 ns/iter (+/- 2,131)
test bench_parse_text      ... bench:         624 ns/iter (+/- 43)
test bench_parse_variable  ... bench:       2,600 ns/iter (+/- 221)
test bench_render_template ... bench:       7,817 ns/iter (+/- 554)
test bench_render_text     ... bench:         262 ns/iter (+/- 18)
test bench_render_variable ... bench:       1,392 ns/iter (+/- 106)

Looks like it took about 200ns off of both `parse` and `render`.

Fixes #188

BREAKING CHANGE: `Context` now takes in an `Arc` to the filters.

- [ ] Tests created for any new feature or regression tests for bugfixes.